pyproject.toml
[tool.poetry]
name = "koa-bot"
version = "0.30.5"
description = ""
authors = []
readme = "README.md"
packages = [
    { include = "koabot" },
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Natural Language :: English",
    "Operating System :: Microsoft :: Windows",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.10",
    "Topic :: Internet",
]


[tool.poetry.dependencies]
python = "^3.10"
appdirs = "^1.4.4"
BASC-py4chan = "^0.6.6"
commentjson = "^0.9.0"
"discord.py" = {git = "https://github.com/Rapptz/discord.py", extras = ["voice", "speed"], rev="36f039a"}
emoji = "^1.7.0"
forex-python = "^1.8"
ImageHash = "^4.2.1"
mergedeep = "^1.3.4"
num2words = "^0.5.10"
Pint = "^0.18"
PixivPy-Async = {extras = ["speedups"], version = "^1.2.14"}
pymediawiki = "^0.7.1"
pytz = "^2022.1"
tldextract = "^3.3.0"
tweepy = {extras = ["async"], version = "^4.10.0"}
single-source = "^0.3.0"
yt-dlp = "^2022.3.8"
uvloop = {platform = "linux", version = "*", optional = true}
aiodns = {version = "*", optional = true}
cchardet = {version = "*", optional = true}
Brotli = {version = "*", optional = true}
asyncpraw = "^7.5.0"
thefuzz = {extras = ["speedup"], version = "^0.19.0"}
tqdm = "^4.64.0"
dataclass-wizard = "^0.22.0"
aiosqlite = "^0.17.0"

[tool.poetry.dev-dependencies]
autopep8 = "^1.6.0"
pylint = "2.12.1"
rope = "^0.23.0"
wheel = "^0.37.1"

[tool.poetry.extras]
speed = ["uvloop","aiodns","cchardet","Brotli"]

[tool.autopep8]
max-line-length = 120

[tool.isort]
sections = ["FUTURE", "STDLIB", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]

# https://pylint.pycqa.org/en/latest/user_guide/message-control.html
    # [C] convention related checks
    # [R] refactoring related checks
    # [W] various warnings
    # [E] errors, for probable bugs in the code
    # [F] fatal, if an error occurred which prevented pylint from doing further processing.
[tool.pylint.messages_control]
max-line-length = 120
good-names-rgxs= "^[_a-z][_a-z0-9]?$"
disable = [
#    "anomalous-backslash-in-string",   # W1401
#    "line-too-long",                   # C0301
#    "invalid-name",                    # C0103
    "assigning-non-slot",               # E0237
    "no-self-use",                      # R0201
]

[tool.pylint.basic]
docstring-min-length=10
no-docstring-rgx="(?:^_|main$)"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
